+++ Fun.Blazor.Docs.Server/CustomElementsDemo.fs
@@ -1,5 +1,6 @@
 namespace Fun.Blazor.Docs.Server
 
+open System
 open Microsoft.AspNetCore.Components
 open Fun.Blazor
 open Fun.Htmx
@@ -14,10 +15,16 @@ type ServerDemoCounter() as this =
     member val count_from_query = 0 with get, set
 
     [<Parameter>]
-    member val count_from_form = 0 with get, set
+    member val count_from_form = "" with get, set
+    
+    [<Parameter>]
+    member val is_loading = false with get, set
+
+    [<Parameter>]
+    member val time = DateTime.Now with get, set
 
     override _.Render() = div {
-        p { $"count = {count}; q {this.count_from_query}; f {this.count_from_form}" }
+        p { $"count = {count}; count_from_query {this.count_from_query}; count_from_form {this.count_from_form}; is_loading {this.is_loading}; time {this.time}" }
         button {
             onclick (fun _ -> count <- count + 1)
             "Click me"