+++ Fun.Blazor/DslElementBuilder.fs
@@ -1,5 +1,6 @@
 ﻿namespace Fun.Blazor
 
+open System.Reflection
 open Microsoft.AspNetCore.Components
 open Operators
 open Internal
@@ -533,8 +534,6 @@ module Elts =
 
     let strong = EltWithChildBuilder "strong"
 
-    let styleElt = EltWithChildBuilder "style"
-
     let sub = EltWithChildBuilder "sub"
 
     let summary = EltWithChildBuilder "summary"
@@ -606,12 +605,6 @@ module Elts =
             index + 1
         )
 
-    let inline stylesheet (x: string) =
-        link {
-            rel "stylesheet"
-            href x
-        }
-
     let inline baseUrl (x: string) = base' { href x }
 
     let inline viewport (x: string) = meta {
@@ -624,37 +617,29 @@ module Elts =
     /// Can be used to build shared dom attributes fragment
     let domAttr = DomAttrBuilder()
 
-    let styleBuilder = StyleBuilder()
-    let cssBuilder = Fun.Css.CssBuilder()
-    let ruleset ruleName = RulesetBuilder ruleName
-    
-    let inline keyframes identifier = KeyFramesBuilder identifier
-    let inline keyframe (x: int) = KeyFrameBuilder(sprintf "%d%%" x)
-    let keyframeFrom = KeyFrameBuilder "from"
-    let keyframeTo = KeyFrameBuilder "to"
 
+    type html with
 
-    /// Build a style string
-    let styleStr = StyleStrBuilder()
+        /// Build hidden inputs for a given state's public properties. 
+        /// If the type is IComponent, only the property with attribute ParameterAttribute will be taken
+        static member createHiddenInputs<'T> (data: 'T) =
+            let ty = typeof<'T>
 
-    /// Short name for StyleBuilder
-    let style = styleBuilder
-
-    /// <summary>
-    /// Short name for cssBuilder
-    /// You can use it as build block when you have complex logic for style
-    /// </summary>
-    /// <example>
-    /// <code lang="fsharp">
-    /// div {
-    ///     style {
-    ///         color "red"
-    ///         if true then
-    ///             css {
-    ///                 backgroundColor "green"
-    ///             }
-    ///     }
-    /// }
-    /// </code>
-    /// </example>
-    let css = cssBuilder
+            data.GetType().GetProperties(BindingFlags.Instance ||| BindingFlags.Public)
+            |> Seq.choose (fun p ->
+                if ty = typeof<IComponent> then
+                    let attr = p.GetCustomAttribute<ParameterAttribute>()
+                    if isNull attr then None else Some p
+                else
+                    Some p
+            )
+            |> Seq.map (fun p ->
+                match p.GetValue data with
+                | null -> html.none
+                | x -> input {
+                    type' InputTypes.hidden
+                    name p.Name
+                    value x
+                  }
+            )
+            |> html.fragment